+++ EZY/Assets.xcassets/calendar.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "calendar.png",
+      "filename" : "Calendar.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
