+++ EZY/Assets.xcassets/clock.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Clock.png",
+      "filename" : "Time Square.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
