+++ EZY/Assets.xcassets/user-3.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Location.svg",
+      "filename" : "3 User.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
