+++ EZY/Assets.xcassets/user.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Location.svg",
+      "filename" : "Profile.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
