+++ EZY/Assets.xcassets/work.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Location.svg",
+      "filename" : "Work (1).svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
