+++ EZY/Controllers/AddToDoViewController.swift
@@ -10,6 +10,7 @@ import SnapKit
 import Then
 import Alamofire
 
+
 class AddToDoViewController:UIViewController{
     
     //MARK: - Properties
@@ -120,7 +121,9 @@ class AddToDoViewController:UIViewController{
     
     @objc func todobackbtn(){
         //전페이지로 되돌아가는 버튼
-        navigationController?.popViewController(animated: true)
+        let vc = LoginViewController()
+        vc.modalPresentationStyle = .fullScreen
+        present(vc, animated: false, completion: nil)
     }
     
     
@@ -218,5 +221,28 @@ class AddToDoViewController:UIViewController{
   
     }
 }
-
+#if DEBUG
+import SwiftUI
+struct ViewControllerRepresentable: UIViewControllerRepresentable {
+    
+func updateUIViewController(_ uiView: UIViewController,context: Context) {
+        // leave this empty
+}
+@available(iOS 13.0.0, *)
+func makeUIViewController(context: Context) -> UIViewController{
+    AddToDoViewController() // 이름 바꾸기
+    }
+}
+@available(iOS 13.0, *)
+struct ViewControllerRepresentable_PreviewProvider: PreviewProvider {
+    static var previews: some View {
+        Group {
+            ViewControllerRepresentable()
+                .ignoresSafeArea()
+                .previewDisplayName(/*@START_MENU_TOKEN@*/"Preview"/*@END_MENU_TOKEN@*/)
+                .previewDevice(PreviewDevice(rawValue: "iPhone 11"))
+        }
+        
+    }
+} #endif
