+++ EZY/Controllers/MoreCalendarModalsViewController.swift
@@ -12,6 +12,7 @@ import SnapKit
 protocol BulletinDelegate: class {
     func onTapClose()
 }
+
 class MoreCalendarModalsViewController : UIViewController{
     
     
@@ -22,7 +23,7 @@ class MoreCalendarModalsViewController : UIViewController{
         $0.backgroundColor = .white
         $0.layer.cornerRadius = 40
     }
-    
+    let transparentView = UIView()
     let TitleLabel = UILabel().then{
         $0.text = "추가 할 항목을 선택해주세요"
         $0.dynamicFont(fontSize: 22, weight: .thin)
@@ -35,33 +36,36 @@ class MoreCalendarModalsViewController : UIViewController{
         $0.setTitleColor(.black, for: .normal)
     }
     
-    
     private let myToDo : MoreCalendarModalsButton = {
-        let viewModel = MyCustomButtonViewModel(title: "나의 할 일", image: UIImage(systemName: "person")!, color: .EZY_BAC8FF)
+        let viewModel = MyCustomButtonViewModel(title: "나의 할 일", image: UIImage(named: "user")!, color: .EZY_BAC8FF)
+        
         let button = MoreCalendarModalsButton(with: viewModel)
-
+        button.addTarget(self, action: #selector(MyTodo), for: .touchUpInside)
         return button
     }()
     
     private let ourToDo : MoreCalendarModalsButton = {
-        let viewModel = MyCustomButtonViewModel(title: "우리의 할 일", image: UIImage(systemName: "person.3")!, color: .EZY_CFCBFF)
+        let viewModel = MyCustomButtonViewModel(title: "우리의 할 일", image: UIImage(named: "user-3")!, color: .EZY_CFCBFF)
         let button = MoreCalendarModalsButton(with: viewModel)
-
+        button.addTarget(self, action: #selector(OurTodo), for: .touchUpInside)
+        
         return button
     }()
     
     private let errand : MoreCalendarModalsButton = {
-        let viewModel = MyCustomButtonViewModel(title: "심부름", image: UIImage(systemName: "person.3")!, color: .EZY_AFADFF)
+        let viewModel = MyCustomButtonViewModel(title: "심부름", image: UIImage(named: "work")!, color: .EZY_AFADFF)
         let button = MoreCalendarModalsButton(with: viewModel)
-
+        button.addTarget(self, action: #selector(Errand), for: .touchUpInside)
         return button
     }()
+    
     static func instance() -> MoreCalendarModalsViewController {
         return MoreCalendarModalsViewController(nibName: nil, bundle: nil).then {
             $0.modalPresentationStyle = .overFullScreen
         }
     }
     //MARK: - Lifecycle
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         configureUI()
@@ -75,16 +79,31 @@ class MoreCalendarModalsViewController : UIViewController{
         dismiss(animated: true, completion: nil)
     }
     
+    @objc func MyTodo(){
+        
+        let vc = AddToDoViewController()
+        vc.modalPresentationStyle = .fullScreen
+        present(vc, animated: true, completion: nil)
+    }
+    @objc func OurTodo(){
+        print("DEBUG : Click OurToDo")
+
+    }
+    @objc func Errand(){
+    }
+    
     //MARK: - HELPERS
+
+    
     func configureUI(){
         view.backgroundColor = .clear
         view.addSubview(bgView)
-        view.addSubview(closeBtn)
+//        view.addSubview(closeBtn)
         view.addSubview(TitleLabel)
         view.addSubview(myToDo)
         view.addSubview(ourToDo)
         view.addSubview(errand)
-        closeBtn.addTarget(self, action: #selector(onTapClose), for: .touchUpInside)
+//        closeBtn.addTarget(self, action: #selector(onTapClose), for: .touchUpInside)
         
         
         
@@ -92,7 +111,7 @@ class MoreCalendarModalsViewController : UIViewController{
             make.left.equalToSuperview()
             make.right.equalToSuperview()
             make.bottom.equalToSuperview().offset(30)
-            make.height.equalToSuperview().dividedBy(1.5)
+            make.height.equalToSuperview().dividedBy(1.6)
         }
         
         TitleLabel.snp.makeConstraints { (make) in
@@ -122,9 +141,7 @@ class MoreCalendarModalsViewController : UIViewController{
         }
         
         
-        closeBtn.snp.makeConstraints { (make) in
-            make.center.equalTo(bgView)
-        }
+        
     }
